Abandoning the tutorial heap interface methods
Retooling the HistoryHeap functions to hold true the interface but
updating them to only work off of the message already stored in the
heap. Added functionality to find the last element in the heap without
modifying the size and Len now returns the number of elements in the
list which are not nil.